feat: install test functions from npm

+++ scripts/build.sh
@@ -13,14 +13,18 @@ if [ -z "${1:-}" ]; then
 fi
 
 exp_dir="experiments/$1/functions"
+exp_config="experiments/$1/experiment.json"
 
-if [[ ! -d $exp_dir ]]; then
+if [[ ! -f $exp_config ]]; then
     echo -e "Invalid experiment name\n" | chalk red
     exit 1
 fi
 
 # Ensure updated dependencies
-npm install
+npm install --ignore-scripts
+
+echo "Installing functions from npm" | chalk cyan
+./scripts/install-functions.js $exp_config $exp_dir
 
 chalk -t "{cyan Building experiment} {cyan.bold $1}"
 
@@ -30,11 +34,18 @@ mkdir $exp_dir/_build
 for d in $exp_dir/*; do
   fname=`basename $d`
   [ "$fname" == '_build' ] && continue
+  if [[ -f "$d/.installed" ]]; then
+    echo "Function already built: ${fname}" | chalk cyan
+    cp $exp_config $d
+    cd $d && zip -r ../_build/${fname}.zip * && cd -
+    rm $d/experiment.json
+    continue
+  fi
   echo "Going to build function: ${fname}" | chalk cyan
   echo "process.env.FAASTERMETRICS_FN_NAME='${fname}';$(cat $d/index.js)" > $d/_index.js
   npx ncc build $d/_index.js -o $d/build
   echo $PKG_JSON > $d/build/package.json
-  cp $exp_dir/../experiment.json $d/build/
+  cp $exp_config $d/build/
   cd $d/build && zip -r ../../_build/${fname}.zip * && cd -
   rm -rf $d/build $d/_index.js
 done
@@ -45,13 +56,19 @@ rm -rf $exp_dir/_build/azure
 mkdir $exp_dir/_build/azure
 
 cp misc/azure/host.json $exp_dir/_build/azure
-cp $exp_dir/../experiment.json $exp_dir/_build/azure
+cp $exp_config $exp_dir/_build/azure
 
-for fname in $(jq -r '.program.functions | with_entries(select(.value.provider == "azure" )) | keys[]' $exp_dir/../experiment.json); do
+for fname in $(jq -r '.program.functions | with_entries(select(.value.provider == "azure" )) | keys[]' $exp_config); do
   echo "Going to build azure function: $fname" | chalk cyan
   d=$exp_dir/$fname
   mkdir $exp_dir/_build/azure/$fname
   cat misc/azure/function.json | jq --argjson fn "\"${fname}/{*path}\"" '.bindings[0].route = $fn' > $exp_dir/_build/azure/$fname/function.json
+  if [[ -f "$d/.installed" ]]; then
+    echo "Function already built: ${fname}" | chalk cyan
+    cp -r $d/* $exp_dir/_build/azure/$fname/
+    continue
+  fi
+  echo "Going to build function: ${fname}" | chalk cyan
   echo "process.env.FAASTERMETRICS_FN_NAME='${fname}';$(cat $d/index.js)" > $d/_index.js
   npx ncc build $d/_index.js -o $exp_dir/_build/azure/$fname
   rm -rf $d/_index.js
